Conflicts:
	archivemailagent/archivemailagent.cpp

+++ archivemailagent/archivemailagent.cpp
@@ -99,7 +99,7 @@ void ArchiveMailAgent::reload()
 
 void ArchiveMailAgent::configure( WId windowId )
 {
-    showConfigureDialog(windowId);
+    showConfigureDialog((qulonglong)windowId);
 }
 
 void ArchiveMailAgent::pause()
